Multilingual ASR recipe for IWSLT'23 Marathi-Hindi

+++ egs2/TEMPLATE/asr1/asr.sh
@@ -41,7 +41,7 @@ expdir=exp           # Directory to save experiments.
 python=python3       # Specify python to execute espnet commands.
 copy_feats_to_dir="" # path to dir that has faster access to the GPU machine.
                      # eg: /tmp or /mnt/ssd, where speech feats will be copied
-
+remove_dups=0        # remove duplicate lines from train.txt lm_train.txt
 # Data preparation related
 local_data_opts= # The options given to local/data.sh.
 
@@ -66,7 +66,11 @@ bpe_input_sentence_size=100000000 # Size of input sentence for BPE.
 bpe_nlsyms=         # non-linguistic symbols list, separated by a comma, for BPE
 bpe_char_cover=1.0  # character coverage when modeling BPE
 token_listdir=      # Path to directory of tokenizer model and vocab.
-
+input_token_list_ftype="token_list"
+                         # Is token list just a list of words or filelist with paths to tokens.txt
+                         # Options are 'token_list' or 'token_flist'. The 'token_flist' option is used to
+                         # train multilingual ASR with lang-specific input/output layers.
+                         # It can be just 'token_list' with shared vocab and prompt based training
 # Ngram model related
 use_ngram=false
 ngram_exp=
@@ -99,6 +103,8 @@ ignore_init_mismatch=false      # Ignore initial mismatch
 feats_normalize=global_mvn # Normalizaton layer type.
 num_splits_asr=1           # Number of splitting for lm corpus.
 multilingual_mode=false    # multilingual mode
+lid=            # monolingual decoding from a multilingual mode - lang-specfic in/out layers
+lid_as_prompt=  # monolingual prompt decoding from a multilingual mode
 
 # Upload model related
 hf_repo=
@@ -151,7 +157,7 @@ asr_text_fold_length=150   # fold_length for text data during ASR training.
 lm_fold_length=150         # fold_length for LM training.
 
 # multilingual specific arguments
-utt2cat=false
+# utt2cat=false
 
 help_message=$(cat << EOF
 Usage: $0 --train-set "<train_set_name>" --valid-set "<valid_set_name>" --test_sets "<test_set_names>"
@@ -199,8 +205,14 @@ Options:
     --bpe_nlsyms              # Non-linguistic symbol list for sentencepiece, separated by a comma. (default="${bpe_nlsyms}").
     --bpe_char_cover          # Character coverage when modeling BPE (default="${bpe_char_cover}").
     --token_listdir           # Path to token_listdir. By default it will be determined
+    --remove_dups             # Remove duplicates from train.txt and lm_train.txt (default="${remove_dups}")
                               # automatically based on above tokenization related options.
                               # (default=)
+    input_token_list_ftype=  # Is token list just a list of words or filelist with paths
+                             # to tokens.txt. Options are 'token_list' or 'token_flist'.
+                             # The 'token_flist' option is used to train multilingual ASR with
+                             # lang-specific input/output layers. It can be just 'token_list' with # shared vocab and prompt based training.
+                             # (default: ${input_token_list_ftype})
 
     # Language model related
     --lm_tag          # Suffix to the result dir for language model training (default="${lm_tag}").
@@ -229,7 +241,10 @@ Options:
     --feats_normalize  # Normalizaton layer type (default="${feats_normalize}").
     --num_splits_asr   # Number of splitting for lm corpus  (default="${num_splits_asr}").
     --multilingual_mode # true or false, (default:false).
-                        # true in case each langauage has specific vocab
+                        # If true, expects utt2category file where category is langauge ID.
+                        # Soft links from {lid.scp, utt2lang }-> utt2category file
+                        # is also required.
+                        # can be used in conjunction with input_token_list_ftype
 
     # Decoding related
     --inference_tag       # Suffix to the result dir for decoding (default="${inference_tag}").
@@ -242,6 +257,11 @@ Options:
     --download_model      # Download a model from Model Zoo and use it for decoding (default="${download_model}").
     --use_streaming       # Whether to use streaming decoding (default="${use_streaming}").
     --use_maskctc         # Whether to use maskctc decoding (default="${use_streaming}").
+    --lid             # monolingual decoding from a multilingual mode - lang-specfic in/out layers
+    --lid_as_prompt   # monolingual prompt decoding from a multilingual mode. Depending on the
+                      # model, you should either lid or lid_as_prompt but not both, because they
+                      # based on different assumptions about the model
+
 
     # [Task dependent] Set the datadir name created by local/data.sh
     --train_set     # Name of training set (required).
@@ -459,7 +479,7 @@ if ! "${skip_data_prep}"; then
     if [ ${stage} -le 1 ] && [ ${stop_stage} -ge 1 ]; then
         log "Stage 1: Data preparation for ${datadir}/${train_set}, ${datadir}/${valid_set}, etc."
         # [Task dependent] Need to create data.sh for new corpus
-        local/data.sh ${local_data_opts}
+        local/data.sh "${local_data_opts}"
     fi
 
     if [ ${stage} -le 2 ] && [ ${stop_stage} -ge 2 ]; then
@@ -496,7 +516,7 @@ if ! "${skip_data_prep}"; then
             # If nothing is need, then format_wav_scp.sh does nothing:
             # i.e. the input file format and rate is same as the output.
 
-            for dset in "${train_set}" "${valid_set}" ${test_sets}; do
+             for dset in "${train_set}" "${valid_set}" ${test_sets}; do
                 if [ "${dset}" = "${train_set}" ] || [ "${dset}" = "${valid_set}" ]; then
                     _suf="/org"
                 else
@@ -647,6 +667,14 @@ if ! "${skip_data_prep}"; then
 
         # shellcheck disable=SC2002
         cat ${lm_train_text} | awk ' { if( NF != 1 ) print $0; } ' > "${data_feats}/lm_train.txt"
+
+        if [ "${remove_dups}" == "1" ]; then
+            log "Removing duplicates from ${data_feats}/lm_train.txt.orig"
+            cp -v "${data_feats}/lm_train.txt" "${data_feats}/lm_train.txt".orig
+            cat "${data_feats}/lm_train.txt.orig" | sort -u > "${data_feats}/lm_train.txt"
+        fi
+
+
     fi
 
 
@@ -656,12 +684,28 @@ if ! "${skip_data_prep}"; then
 
             mkdir -p "${bpedir}"
             # shellcheck disable=SC2002
-            cat ${bpe_train_text} | cut -f 2- -d" "  > "${bpedir}"/train.txt
+
+            if [ -f ${bpedir}/train.txt ]; then
+                echo "INFO: ${bpedir}/train.txt already exists. Not overwriting."
+            else
+                cat ${bpe_train_text} | cut -f 2- -d" "  > "${bpedir}"/train.txt
+            fi
+
+            if [ "${remove_dups}" == "1" ]; then
+                log "Removing duplicates from ${bpedir}/train.txt"
+                cp -v ${bpedir}/train.txt ${bpedir}/train.txt.orig
+                cat ${bpedir}/train.txt.orig | sort -u > ${bpedir}/train.txt
+            fi
 
             if [ -n "${bpe_nlsyms}" ]; then
                 _opts_spm="--user_defined_symbols=${bpe_nlsyms}"
             else
                 _opts_spm=""
+            fi
+
+            if [ -f "${bpeprefix}.model" ]; then
+                echo "${bpeprefix}.model already exists. Choose a different directory or skip this step"
+                exit 1;
             fi
 
             spm_train \
@@ -728,7 +772,7 @@ fi
 # create utt2category symlink to lid.scp to create batches of same language during training
 #if [ ${utt2cat} = true ]; then
 #    python3 lid_scp.py --bpedir ${bpedir} \
-#                       --dump_dirs ${data_feats}/${train_set} ${data_feats}/${valid_set} ${data_feats}/${test_sets} 
+#                       --dump_dirs ${data_feats}/${train_set} ${data_feats}/${valid_set} ${data_feats}/${test_sets}
 #if
 
 # ========================== Data preparation is done here. ==========================
@@ -737,7 +781,7 @@ fi
 if ! "${skip_train}"; then
     if "${use_lm}"; then
         if [ ${stage} -le 6 ] && [ ${stop_stage} -ge 6 ]; then
-            log "Stage 6: LM collect stats: train_set=${data_feats}/lm_train.txt, dev_set=${lm_dev_text}"
+            log "Stage 6: LM collect stats: train_set=${lm_train_text}, dev_set=${lm_dev_text}"
 
             _opts=
             if [ -n "${lm_config}" ]; then
@@ -750,9 +794,9 @@ if ! "${skip_train}"; then
             _logdir="${lm_stats_dir}/logdir"
             mkdir -p "${_logdir}"
             # Get the minimum number among ${nj} and the number lines of input files
-            _nj=$(min "${nj}" "$(<${data_feats}/lm_train.txt wc -l)" "$(<${lm_dev_text} wc -l)")
+            _nj=$(min "${nj}" "$(<${lm_train_text} wc -l)" "$(<${lm_dev_text} wc -l)")
 
-            key_file="${data_feats}/lm_train.txt"
+            key_file="${lm_train_text}"
             split_scps=""
             for n in $(seq ${_nj}); do
                 split_scps+=" ${_logdir}/train.${n}.scp"
@@ -777,7 +821,7 @@ if ! "${skip_train}"; then
             # NOTE: --*_shape_file doesn't require length information if --batch_type=unsorted,
             #       but it's used only for deciding the sample ids.
             # shellcheck disable=SC2086
-            ${train_cmd} JOB=1:"${_nj}" "${_logdir}"/stats.JOB.log \
+            ${cpu_cmd} JOB=1:"${_nj}" "${_logdir}"/stats.JOB.log \
                 ${python} -m espnet2.bin.lm_train \
                     --collect_stats true \
                     --use_preprocessor true \
@@ -787,7 +831,7 @@ if ! "${skip_train}"; then
                     --non_linguistic_symbols "${nlsyms_txt}" \
                     --cleaner "${cleaner}" \
                     --g2p "${g2p}" \
-                    --train_data_path_and_name_and_type "${data_feats}/lm_train.txt,text,text" \
+                    --train_data_path_and_name_and_type "${lm_train_text},text,text" \
                     --valid_data_path_and_name_and_type "${lm_dev_text},text,text" \
                     --train_shape_file "${_logdir}/train.JOB.scp" \
                     --valid_shape_file "${_logdir}/dev.JOB.scp" \
@@ -814,7 +858,7 @@ if ! "${skip_train}"; then
 
 
         if [ ${stage} -le 7 ] && [ ${stop_stage} -ge 7 ]; then
-            log "Stage 7: LM Training: train_set=${data_feats}/lm_train.txt, dev_set=${lm_dev_text}"
+            log "Stage 7: LM Training: train_set=${lm_train_text}, dev_set=${lm_dev_text}"
 
             _opts=
             if [ -n "${lm_config}" ]; then
@@ -832,7 +876,7 @@ if ! "${skip_train}"; then
                 if [ ! -f "${_split_dir}/.done" ]; then
                     rm -f "${_split_dir}/.done"
                     ${python} -m espnet2.bin.split_scps \
-                      --scps "${data_feats}/lm_train.txt" "${lm_stats_dir}/train/text_shape.${lm_token_type}" \
+                      --scps "${lm_train_text}" "${lm_stats_dir}/train/text_shape.${lm_token_type}" \
                       --num_splits "${num_splits_lm}" \
                       --output_dir "${_split_dir}"
                     touch "${_split_dir}/.done"
@@ -845,7 +889,7 @@ if ! "${skip_train}"; then
                 _opts+="--multiple_iterator true "
 
             else
-                _opts+="--train_data_path_and_name_and_type ${data_feats}/lm_train.txt,text,text "
+                _opts+="--train_data_path_and_name_and_type ${lm_train_text},text,text "
                 _opts+="--train_shape_file ${lm_stats_dir}/train/text_shape.${lm_token_type} "
             fi
 
@@ -864,7 +908,7 @@ if ! "${skip_train}"; then
 
             # shellcheck disable=SC2086
             ${python} -m espnet2.bin.launch \
-                --cmd "${cuda_cmd} --name ${jobname}" \
+                --cmd "${lm_cuda_cmd} --name ${jobname}" \
                 --log "${lm_exp}"/train.log \
                 --ngpu "${ngpu}" \
                 --num_nodes "${num_nodes}" \
@@ -917,7 +961,7 @@ if ! "${skip_train}"; then
     fi
     if [ ${stage} -le 9 ] && [ ${stop_stage} -ge 9 ]; then
         if "${use_ngram}"; then
-            log "Stage 9: Ngram Training: train_set=${data_feats}/lm_train.txt"
+            log "Stage 9: Ngram Training: train_set=${lm_train_txt}"
             cut -f 2- -d " " ${data_feats}/lm_train.txt | lmplz -S "20%" --discount_fallback -o ${ngram_num} - >${ngram_exp}/${ngram_num}gram.arpa
             build_binary -s ${ngram_exp}/${ngram_num}gram.arpa ${ngram_exp}/${ngram_num}gram.bin
         else
@@ -982,7 +1026,7 @@ if ! "${skip_train}"; then
             # need extra options for collect_stats
             _opts+="--train_data_path_and_name_and_type ${_asr_train_dir}/lid.scp,lid,text "
             _opts+="--valid_data_path_and_name_and_type ${_asr_valid_dir}/lid.scp,lid,text "
-            _opts+="--input_token_list_ftype token_flist "
+            _opts+="--input_token_list_ftype ${input_token_list_ftype} "
         fi
 
         # 2. Generate run.sh
@@ -1241,6 +1285,13 @@ if ! "${skip_eval}"; then
           fi
         fi
 
+        if [ -n "${lid}" ]; then
+            _opts+="--lid ${lid}"
+        fi
+        if [ -n "${lid_as_prompt}" ]; then
+            _opts+="--lid_as_prompt ${lid_as_prompt}"
+        fi
+
         for dset in ${test_sets}; do
             _data="${data_feats}/${dset}"
             _dir="${asr_exp}/${inference_tag}/${dset}"
@@ -1312,6 +1363,14 @@ if ! "${skip_eval}"; then
         for dset in ${test_sets}; do
             _data="${data_feats}/${dset}"
             _dir="${asr_exp}/${inference_tag}/${dset}"
+
+            if [ ! -z ${lid_as_prompt} ]; then
+                # removes lang id (eg: [cs] or [pl] or [en]) from the hyp text
+                # so that WER, CER, TER can be computed reliably
+                ${python} pyscripts/utils/remove_lid_prompt.py \
+                    -text ${_dir}/text \
+                    -lid_prompt ${lid_as_prompt}
+            fi
 
             for _type in cer wer ter; do
                 [ "${_type}" = ter ] && [ ! -f "${bpemodel}" ] && continue
